Fix check for pointer compression.
The `CXXFLAGS` were not passed to the compile of the test executable,
which causes issues if some flags in `LDFLAGS` depend on flags also
being in `CXXFLAGS`.

Also, the exit status of the compilation is not checked, so this could
continue to guess a result even if there's no test executable.